fix: link on header landing page

+++ src/apis/admin.ts
@@ -30,7 +30,7 @@ const getListDeletedUsers = async () => {
 
 const deleteUser = async (id: string | number) => {
      try {
-          return await axiosClient.delete(`users/${id}`)
+          return await axiosClient.delete(`admin/users/${id}`)
      } catch (error) {
           console.error("Error deleting user:", error);
           throw error;
